Refactor: Extract shared geographic utilities to eliminate code duplication
✅ Copilot Code Review Fixes:
- Create shared @/lib/geoUtils.ts with calculateDistance, deg2rad, and constants
- Remove duplicated Haversine formula from Home.tsx and barcelonaMockData.ts
- Centralize BARCELONA_CENTER constant for consistency
- Add distanceFromBarcelonaCenter helper function

🔧 Code Quality Improvements:
- DRY principle: Single source of truth for geographic calculations
- Better maintainability: Changes only need to be made in one place
- Type safety: Consistent function signatures across modules
- Documentation: Proper JSDoc comments for utility functions

📊 Impact:
- Eliminated ~40 lines of duplicated code
- Consistent Barcelona coordinate references across app
- Easier testing and debugging of distance calculations
- Follows established @/lib pattern for shared utilities

Thanks to Copilot code review for catching the duplication!
