+++ src/pages/Home.tsx
@@ -10,25 +10,7 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { SliderPicker } from "@/components/SliderPicker";
 import { RefreshCw } from "lucide-react";
 import { toast } from "sonner";
-
-// Calculate distance between two points in km using Haversine formula
-const calculateDistance = (coords1: [number, number], coords2: [number, number]): number => {
-  const R = 6371; // Earth's radius in km
-  const dLat = deg2rad(coords2[1] - coords1[1]);
-  const dLon = deg2rad(coords2[0] - coords1[0]);
-  
-  const a =
-    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-    Math.cos(deg2rad(coords1[1])) * Math.cos(deg2rad(coords2[1])) *
-    Math.sin(dLon / 2) * Math.sin(dLon / 2);
-  
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return R * c; // Distance in km
-};
-
-const deg2rad = (deg: number): number => {
-  return deg * (Math.PI / 180);
-};
+import { calculateDistance, BARCELONA_CENTER } from "@/lib/geoUtils";
 
 const Home = () => {
   const [viewMode, setViewMode] = useState<"map" | "list">("map");
@@ -45,14 +27,11 @@ const Home = () => {
   const processedTables = useMemo(() => {
     if (!tables) return [];
     
-    // Barcelona city center (Plaça de Catalunya)
-    const userLocation: [number, number] = [2.1700, 41.3874];
-    
     return tables.map(table => {
       // Calculate distance from Barcelona city center
       const distance = calculateDistance(
-        userLocation,
-        table.location ? table.location.coordinates : [2.1700, 41.3874]
+        BARCELONA_CENTER,
+        table.location ? table.location.coordinates : BARCELONA_CENTER
       );
       
       return {
